+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [0.75.0] 2024-05-22
+
+- Added Timeline entity
+- Refactored Pantheon entity
+
 ## [0.74.1] 2024-05-22
 
 - Added Rank attribute to Person entity